Added File rest calls
Updataed db and module to use GUIDs instead of ints

This is bacause in the newer versions, we have no way of authenticating
get requests on files. CKeditor doesnt allow uss to add those.

I didnt want to use a cookies.

So we are using guids instead.

If someones shares the file it can be downloaded until it is deleted.

This is so you won't be able to just sequentiall get all the files